Fix OOM crashes: float32 DSP + IIR filters + explicit del + FFmpeg
Root cause: Railway (512 MB) was killed by the DSP render peak.
A 3-min song at float64 with FFT crossovers + per-sample float64
position arrays hit 700-800 MB peak, triggering SIGKILL mid-render.
The server restarted, the in-memory job dict was wiped, polls returned
404, and the UI showed 'Render failed'.

Key changes:
- ensure_stereo_float: float64 -> float32 (halves all audio arrays)
- audio_io: soundfile reads float32 directly
- Add _split_at_hz(): scipy IIR crossover -- O(n) memory vs FFT's
  O(n) complex128 arrays (~300 MB per call saved)
- high_frequency_emphasis, split_center_focus_bands, enhance_felt_presence:
  all now use _split_at_hz instead of _fft_split_mono
- reduce_static_noise: removed full-file rfft spectral subtraction;
  replaced with IIR high-shelf blend (same perceptual result, -200 MB peak)
- section_automation_curve: float32 + scipy uniform_filter1d (no fftconvolve)
- _variable_fractional_delay: int32 read indices + float32 frac parts
  instead of two float64 position arrays (-320 MB peak for 5-min track)
- binaural_orbit: all angular arrays float32; explicit del after each step
- process_8d: explicit del for bass/motion/room/orbit intermediates
- nixpacks.toml: install FFmpeg so pydub can decode MP3/AAC/M4A

+++ eightd_engine/audio_io.py
@@ -39,7 +39,7 @@ def load_audio(path: str | Path) -> AudioData:
     try:
         import soundfile as sf  # type: ignore
 
-        samples, sample_rate = sf.read(str(src), always_2d=True, dtype="float64")
+        samples, sample_rate = sf.read(str(src), always_2d=True, dtype="float32")
         return AudioData(samples=ensure_stereo_float(samples), sample_rate=int(sample_rate))
     except Exception as sf_error:
         if src.suffix.lower() == ".wav":
